requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.3.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
distro==1.8.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
ffpyplayer==4.5.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
llvmlite==0.40.1
lxml==4.9.2
MarkupSafe==2.1.3
more-itertools==10.0.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
openai==0.27.4
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydub==0.25.1
pyparsing==3.1.0
PyPDF2==3.0.1
pyTelegramBotAPI==4.10.0
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
sniffio==1.3.0
sympy==1.12
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
zipp==3.16.2
